requirements.txt
lxml
requests
requests_oauthlib
pillow
BeautifulSoup4
jsmin